begin project restructuring
