+++ src/BrowserRouter.jsx
@@ -1,16 +1,16 @@
 import {createBrowserRouter, RouterProvider} from "react-router-dom";
-import HomePage from './pages/HomePage.jsx'
+import HomePage from './pages/HomePage/HomePage.jsx'
 import AboutPage from './pages/AboutPage.jsx'
-import CarsPage from './pages/CarsPage.jsx'
-import LoginPage from './pages/LoginPage.jsx'
-import SignupPage from "./pages/SignupPage.jsx";
+import CarsPage from './pages/CarsPage/CarsPage.jsx'
+import LoginPage from './pages/LoginPage/LoginPage.jsx'
+import SignupPage from "./pages/SignupPage/SignupPage.jsx";
 import ReviewPage from "./pages/ReviewPage.jsx";
-import MyPage from "./pages/MyPage.jsx";
+import MyPage from "./pages/MyPage/MyPage.jsx";
 import NotFoundPage from "./pages/NotFoundPage.jsx";
-import UserPage from "./pages/UserPage.jsx";
+import UserPage from "./pages/UserPage/UserPage.jsx";
 import ManageOwnedCarsPage from "./pages/ManageOwnedCarsPage.jsx";
 import SettingsPage from "./pages/SettingsPage.jsx";
-import UsersPage from "./pages/UsersPage.jsx";
+import UsersPage from "./pages/UsersPage/UsersPage.jsx";
 import OrdersPage from "./pages/OrdersPage.jsx";
 import AddNewCar from "./components/AddNewCar.jsx";
 import App from "./App.jsx";
