+++ src/components/Car/Feature.jsx
@@ -1,6 +1,3 @@
-import "../static/css/car.css";
-import PropTypes from 'prop-types';
-
 
 function Feature ({feature}) {
     return (
