+++ src/components/OrderList/OrderList.jsx
@@ -1,7 +1,7 @@
-import SearchableFieldTable from "./SearchableFieldTable.jsx";
+import SearchableFieldTable from "../../components/SearchableFieldTable/SearchableFieldTable.jsx";
 import {useNavigate} from "react-router-dom";
 import {useEffect, useState} from "react";
-import {fetchWithAuth} from "../static/js/auth.js";
+import {fetchWithAuth} from "../../static/js/auth.js";
 
 function ReviewCustomer(order, navigate) {
     return (
