+++ src/pages/LoginPage/LoginPage.jsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 import {Link, useNavigate} from 'react-router-dom';
-import useTitle from "../components/useTitle.jsx";
-import "../static/css/loginAndSignup.css";
-import Conflict from "./ConflictPage.jsx";
+import useTitle from "../../components/useTitle.jsx";
+import "../../static/css/loginAndSignup.css";
+import Conflict from "../ConflictPage.jsx";
