+++ src/pages/OrdersPage.jsx
@@ -1,5 +1,5 @@
 //TODO: Fetch the orders from the database and display them in a table
-import OrderList from "../components/OrderList.jsx";
+import OrderList from "../components/OrderList/OrderList.jsx";
 import useTitle from "../components/useTitle.jsx";
 import React, { useState, useEffect } from 'react';
 import {fetchWithAuth} from "../static/js/auth.js";
