+++ src/static/css/loginAndSignup.css
@@ -1,4 +1,3 @@
-@import "./index.css";
 
 .login{
     display: flex;
